+++ R/utils.R
@@ -94,7 +94,8 @@ hdf5_matrix <- function(datapath, dataset, start = 0, end = NULL, normalizer = N
 #' Normalize a matrix or nd-array
 #' 
 #' @param x Matrix or array to normalize
-#' @param axis Axis along which to normalize
+#' @param axis Axis along which to normalize. Axis indexes are 1-based
+#'   (pass -1 to select the last axis).
 #' @param order Normalization order (e.g. 2 for L2 norm) 
 #' 
 #' @return A normalized copy of the array.
@@ -103,7 +104,7 @@ hdf5_matrix <- function(datapath, dataset, start = 0, end = NULL, normalizer = N
 normalize <- function(x, axis = -1, order = 2) {
   keras$utils$normalize(
     x = x,
-    axis = as.integer(axis),
+    axis = as_axis(axis),
     order = as.integer(order)
   )
 }